Avoid `.always_inline` to improve debugging experience.
The `SystemImpl` wrapper used by convenience functions such as `ADD_SYSTEM` call the user's system update function (`fn_system`) via `@call(.always_inline, ...`. In theory there is no reason not to use `.always_inline` here, but in practice it results in a worse debugging experience as of Zig 0.14.0. For example, on Windows crashes inside `fn_system` are attributed to this `@call`, and it is not possible to set breakpoints on code inside `fn_system`.